refactor: 매퍼영역에 있던 DTO DTO 디렉터리로 옮김
